% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bashR-package.R
\docType{package}
\name{bashR-package}
\alias{bashR}
\alias{bashR-package}
\title{bashR: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
