% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_economic_assumptions.R
\name{clean_price_inflation}
\alias{clean_price_inflation}
\title{Clean ONS Price data}
\usage{
clean_price_inflation(raw_price_inflation)
}
\arguments{
\item{raw_price_inflation}{tibble of ONS price inflation data}
}
\value{
tibble with adjustments from the description
}
\description{
1. Set first column name to `month`
2. Create a `date` field, set to 1st of month
3. Rename `cpi = D7G7`, `rpi = CZBH`
}
\examples{
raw_price_inflation <- read_raw_price_inflation("data/consumerpriceinflationdetailedreferencetables.xlsx")
clean_price_inflation(raw_price_inflation)
}
