% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_economic_assumptions.R
\name{download_price_inflation}
\alias{download_price_inflation}
\title{Download, save, and read}
\usage{
download_price_inflation(
  link = get_price_inflation_source()$link,
  file = paste("data", basename(link), sep = "/")
)
}
\arguments{
\item{link}{a string of the link to download}

\item{file}{a string of the location to save the downloaded dataset}
}
\value{
An (invisible) integer code, 0 for success and non-zero for failure
}
\description{
A wrapper around `download.file()`, with defaults using `[get_price_inflation_source]()`, to download and save price inflation data from the ONS.
}
\examples{
download_price_inflation()
}
