% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all.R
\name{get_all}
\alias{get_all}
\title{Extract All Functions and Variables from a Namespace}
\usage{
get_all(package, envir = parent.frame())
}
\arguments{
\item{package}{a symbol or string}

\item{envir}{an environment to import values into. E.g., `new.env()`. Defaults to `parent.frame()`.}
}
\value{
`envir` with items from `package` imported or overwritten.
}
\description{
This function will enable the user to access all functions / variables within a Namespace.
E.g., the user could use it to access private functions from a package.
}
\details{
By default, all items are imported into the parent environment.
E.g., values will be imported into the Global Environment if the user calls this function,
}
\examples{
get_all(dplyr); mutate.data.frame
dplyr <- get_all(dplyr, rlang::env()); dplyr$mutate.data.frame
}
