% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_economic_assumptions.R
\name{get_ons_deaths_source}
\alias{get_ons_deaths_source}
\title{Links to ONS Death datasets}
\usage{
get_ons_deaths_source(month, year)
}
\arguments{
\item{month}{string of month data is required. Months are lowercase in English. E.g., "january".}

\item{year}{integer of year data is required. E.g., 202.}
}
\value{
named list of online addresses to the ONS site and source of Death data
}
\description{
Links to ONS Death datasets
}
\examples{
get_ons_deaths_source("january", 2024)
}
