% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_economic_assumptions.R
\name{get_price_inflation_source}
\alias{get_price_inflation_source}
\title{Links to ONS Price Inflation datasets}
\usage{
get_price_inflation_source()
}
\value{
named list of online addresses to the ONS site and source of price inflation data
}
\description{
Links to ONS Price Inflation datasets
}
\examples{
get_price_inflation_source()
}
