% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_economic_assumptions.R
\name{read_raw_price_inflation}
\alias{read_raw_price_inflation}
\title{Extract inflation information, save tidy format}
\usage{
read_raw_price_inflation(
  price_inflation_file,
  sheet = "Table 1",
  range = "B14:H51"
)
}
\arguments{
\item{price_inflation_file}{a string of the price inflation file}

\item{sheet}{a string of the sheet to read}

\item{range}{a string of the range to read}
}
\value{
a tibble
}
\description{
A wrapper around `readxl::read_xlsx()` to read ONS Price Inflation into R.
}
\examples{
# Download the dataset
# link <- get_price_inflation_source()$link
# file <- paste("data", basename(link), sep = "/")
# download_price_inflation(link, file)
# read_raw_price_inflation(file)
}
