% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sudo}
\alias{sudo}
\title{sudo$}
\usage{
sudo(cmd, intern = F, ignore.stdout = F, ignore.stderr = F)
}
\arguments{
\item{cmd}{A bash command to execute as sudo}
}
\description{
This function executes the given command as sudo
}
